requirements.txt
blinker==1.4
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
pypng==0.0.20
PyQRCode==1.2.1
pyzbar==0.1.8
SQLAlchemy==1.4.21
Werkzeug==2.0.1
